Added a new API to the vtkArcSource, that is consistent and unambiguous
The original API, which remains the default, allows for ambiguous
specifications (when center, pt1, and pt2 are aligned, thus not
defining properly the plane of the arc) or even inconsistent ones
(when pt1 and pt2 are not equidistant from center). The goal of
the new API is to fix that.

Change-Id: I78c51f33bbe5f8b9641667ebde8c08d9c89ee56f