Add single, double, and data array functor dispatch classes.
The dispatcher provides the ability to do multiple dispatching based
on a runtime polymorphism. This allows mapping to compile time
polymorphism.
The double dispatcher allows you to call a functor
based on the types of two objects allowing you to call specialized
implementations.
The data array dispatch is a replacement for the vtkTemplateMacro
offering an object based way that is far easier to read and use.

Change-Id: If281ddc1ac803c0ed234e33c9cc0bf4c090fca71