COMP: Fix segfault in vtkWrapJava on underscores in method names.
On OS X 10.5, PowerPC, it was reported that vtkWrapJava crashed
on some input files, and a stack trace identified a strncat call
in part of the code that is meant to handle method names that
contain underscores.  I could not reproduce the problem on my
system (not even by adding asserts to guard against overflows),
so I have re-written the code to eliminate strcat and strchr.

Change-Id: I95b9b48c9ca2ecf0cbefee8b6a13cb91907ce9b3