Fix vtk_module_dep_includes to scan dependencies correctly.
Although vtk_module_dep_includes() is expected to recursively scan the
dependencies to build the include dirs list, it didn't work as expected with the
macro was called repeatedly within the same scope for different modules.  The
problem was that the code that managed the recursion traversal
(_vtk_module_config_recurse) uses variables to avoid duplicate traversal and
that was resulting identifying incorrect duplicate visits when
vtk_module_dep_includes() was used more than once. Fixed that by ensuring that
vtk_module_dep_includes() scope the variable namespace using the modules name.

Change-Id: I23fb97a9dafaf43b7ebdf0f837ee130b3ded9165