vtkfreetype: Include freetype headers robustly
Remove vtkfreetype/include from the module interface's include path.
Use the following shell code to add vtkfreetype/include explicitly to
header inclusions in the freetype public interface.

 paths='
  include/freetype/*.h
  include/freetype/config/*.h
  builds/unix/ftconfig.h.in
  builds/win32/freetype/config/ftoption.h
 '
 grep -l '#include <ft2build\.h>' $paths |
  xargs sed -i 's|include <ft2build\.h>|include <vtkfreetype/include/ft2build.h>|'
 grep -l -R 'define.*<freetype/' include builds |
  xargs sed -i 's|\(define.*\)<freetype/|\1<vtkfreetype/include/freetype/|'

This ensures that the proper headers are included in either a VTK build
tree or a VTK install tree and never conflict with system-installed
freetype headers.

Change-Id: Ic4c63dd37816cdfdabc7d6f79eb6c7ae1c53748d