Do not set (SO|)VERSION properties on VTK executables
The vtk_target_name macro sets the VERSION and SOVERSION properties to
"1" and puts the VTK version number (API version) directly in the
library file name via the OUTPUT_NAME property.  This was written for
VTK libraries and works well for them.  However, since commit 8b0cf94e
(Use the vtk_target macro to export/install wrappers, 2012-06-04) the
vtk_target macro is used also for wrapper executables.  We do not want
or need the "-1" extension added by the VERSION property on executable
targets, so do not set it on them.

Change-Id: I2f20e06c771d2e20f72307566c0c5950c7a96c9a