Fixes bugs when BUILD_SHARED_LIBS was ON.
vtkPythonWrapping was setting dependencies even when BUILD_SHARED_LIBS was ON.
This was resulting in vtkpython dependending on the modules are compile time
acceidentally. Fixed that.

Modules were being added using vtk_add_library() call. That call sets properties
on the target that are not necessarily applicable to "Modules" and was raising
compiler errors when BUILD_SHARED_LIBS was ON. Fixed that. Now vtk_add_library()
is only called when building static (it's needed in that case since the static
module needs to be treated as a regular library and must be added the exports so
that applications can link against it).

Change-Id: Ic0219e7b28d644e9a755fb5f494cbfc6d71dfd09