Adding add_test_tcl() macro.
This macro lets one add a tcl test. It takes the name of the file used in the
test as the first argument and an optional base image directory as the second
argument. However this general macro may not suit all tests in which case a
specifiy cmake code may be required.

Example:

add_test_tcl(mcubes Graphics)

Change-Id: Ie8ab5d344c3fa0bfc880bf637fa77a08b8c34dae