ENH: Remodel the parser rules to better approximate the C++ standard.
Rewrite the parser rules to approximate the ISO 1998 C++.  Several
documents were referenced, primarily the ANSI/ISO 1996 C++ Draft
Standard (the BNF description of the grammar).  The parser remains,
primarily, an empirical implementation of the C++ grammar but these
changes will allow the wrapper tools to handle a greater variety
of input code.

This newly-wrapped method declaration in vtkVector.h had to be changed:
 vtkVector3<T> Cross(const vtkVector3& other) const
to
 vtkVector3<T> Cross(const vtkVector<T>3& other) const
because the wrappers can still not perform context-sensitive name
lookups for template arguments.

Change-Id: Iccdaa3b9f06bb165a79b9a2d8ae99a97b822718b