COMP: Ambiguity error for fabs in vtkTuple::Compare with ints.
The vtkTuple::Compare() method uses fabs(), which depending on the
compiler, might be the C library fabs(), which always takes double,
or the C++ library std::fabs(), which is overloaded.  The latter
takes float, double, and long double... but any other types give
rise to an ambiguous conversion.  This is fixed by replacing fabs()
with std::abs(), which works for integers and floating-point, as
long as both cstdlib and cmath are included.

This problem became visible because a recent wrapper-enhancement
commit caused vtkTuple::Compare to be wrapped for the first time.
This caused the Compare method to be exercised for integer types.

Change-Id: I4816834ae03dc92ca1354d83daa0b99da9c15501