COMP: Use uintptr_t rather than rely on VTK_SIZEOF_VOID_P.
Some VTK code was using VTK_SIZEOF_VOID_P to guess which integer
type was the size of a pointer.  The pointer-sized type uintptr_t
is a cleaner alternative.  It is part of the C99 standard but is
already used in VTK elsewhere, so I believe it should be safe to
use here.

Change-Id: I9903c6165acd0192ce89b1a374d428a59d29246f