Fix search path for vtk module in the build directory.
vtkPythonAppInit searches for the "vtk" module in the build directory at known
locations. When vtkpython was built from ParaView, it ended up adding an
incorrect path since ParaView also contains the Wrapping/Python directory, but
that doesn't include the VTK module. Fixing it by ensure we search for the path
upto "vtk" module and not just the directory containing it.

Change-Id: I1fa684a8e6ec28fecb3103c34dd55974291e773f