Rendering fixes for vtkMathTextActor.
Now `vtkMathTextActor` inherits from `vtkTextActor`
from the `vtkRenderingFreeType` kit. It uses the `vtkTextActor`
to render as a fallback when a `vtkMathTextUtilities`
instance is not available.

Also, the bounding box is computed properly even before
a render (by assuming a display resolution in order
to decide on a texture image size).

Change-Id: I7f9e41f7714cc4e6e0efb42091dc313e0336fa9d