Add GetBoundingBox and use rotations in MathTextUtils.
This patch adds a GetBoundingBox function to vtkMathTextUtilities
and applies rotations from the supplied vtkTextProperty.

The matplotlib MathText interface doesn't have an entry point to
just get the bounding box, so the image is rendered by python as normal,
and then the dimensions are returned.

The MathText library doesn't support rendering rotated text directly,
so a vtkImageReslice filter is used to rotate the image.

Change-Id: I65be970af47c7f8de65f6279c922b6b001f62780