BUG: InteractorStyleImage not getting image for Window/Level.
If an observer has not been set for the StartWindowLevelEvent, then
the vtkInteractorStyleImage is supposed to automatically locate the
last vtkImageActor or vtkImageSlice that was added to the renderer,
and apply the window/level to that actor.  However, there was a bug
where if a non-image actor was added to the renderer after the image
actor, the image actor would not be found.  This patch fixes that bug.

Backport of 022d847a from master to release.

Change-Id: I58baac006ffe66ef367e196c10f98e60ee6e270e