vktDataObjectTree: Fix CopyStructure bug
When the input to CopyStructure() is vtkDataObject, instead of just
bailing out, create a flat tree with same number of blocks as the
source.

This fixes a set of ParaView dashboard tests,
e.g. pvcs.CTHAMRBaseline, pvcs.TestIsoVolume, which all invoke the
function in vtkCompositeDataPipeline to compute multiblock poly data
over amr that contain null blocks.

Change-Id: Ia8696f10fba0d975009adde0419aa4506ef055a2