Change the behavior of vtkDataObjecTree::CopyStructure
When the source iterator is not vtkDataObjecTreeIterator, create a tree
of two levels. The first level is single node of MultiPieceDataSet and
all actual data sets live on the second level. The change is mainly to
make sure that GUI on ParaView do not get overwhelmed by large number
of data sets.

- SetDataSet and GetDataSet are also changed to be consistent.

- The single MultiPieceDataSet node is named "All Blocks" so it
  doesn't look confusing on the GUI.

- A test is created that checks the new behavior.

Change-Id: Ic926242fcfbaa1548b610e3e7e648b354dc8583c