vtkDispatcher: Simplify Add<> interface to a single method
The IBM XL compiler cannot choose between the two overloads of the Add<>
method template, producing errors such as:

 1540-0219 (S) The call to "Add<...>" has no best match.

Work around the problem by simplifying the interface to have only a
single Add<> method that internally calls one of two private
AddInternal<> method templates.  Pass an extra "1" argument to select
between "int" and "long" conversions to help the compiler select the
proper overload.  It will choose the int->long conversion of the second
argument only if it cannot match the pointer form of the first argument.

Change-Id: I2f705761526353c0131160df51028365a0d4a70b