Deprecate vtkColor::Red() getter in favor of GetRed() etc.
The Red(), Green(), Blue() methods don't follow the usual naming pattern.
If future wrappers see SetRed() and GetRed() methods, they can infer
the existence of a property called "Red", but if a method called "Red()"
already exists, this creates an ambiguity that the wrappers will have
to deal with if they do property-based wrapping.

Change-Id: I8a5fa13074512bf7b06e353a5fa7b73efe6ce039