vtkCompositeDataIterator: Move some methods to child class
The two parameters in vtkCompositeDataIterator, TraverseSubTree and
VisitOnlyLeaves, only make sense in the child class
vtkDataObjectTreeIterator, so they are moved there.

For callers, if it just calls VisitOnlyLeavesOn() then we just yank
the call since this is the default behavior anyways; otherwise, we
down cast the iterator to vtkDataObjectIterator in order to call these
functions.

Change-Id: Idbf44426bd1e281219309a824f603bd796b711b1