Fix possible unititialized deltaI array use in vtkWarpTransform.
When inverting a transform via Newton's method, the first iteration
must always compute the derivatives.  The code weakly enforced this
by initializing the lastFunctionValue to VTK_DOUBLE_MAX so that the
current function value would always be less.  This patch adds an
explicit check for "i == 0" where "i" is the iteration counter, to
ensure that the derivatives are always computed on the first iteration.
This issue was discovered by the clang static analyzer.

Change-Id: Ia155fe9134123b1cba4ec50f99092bd79ca085b5