Added SetInputDataObject() API to vtkAlgorithm.
Since setting input data objects is a common practice, making it simpler for
users to set the input by providing API SetInputDataObject(...),
AddInputDataObject(...).

Also, SetInputDataObject(...) has been updated to not modify the Mtime of the
algorithm every time the method is called. Extra logic ensures that the input is
changed only when the data object passed into SetInputDataObject() is changed.

Also added a test that ensures the MTime's work as expected
(TestSetInputDataObject).

Change-Id: I96356e272dd1ff0b9bb1072c2d2ede16f852dfef