Add vtkProteinRibbonFilter class
This adds a new class which generates a protein ribbon (as vtkPolyData)
from the output provided by vtkPDBReader. Currently the entire protein
is rendered as a single spline tube without using the secondary-structure
information (e.g. alpha-helices and beta-sheets). Future work will update
the filter to output a poly-data with the secondary-structure features.

A new test named TestProteinRibbon was added to verify the output of the
filter with the 2LYZ pdb file.

Also, the vtkPDBReader class was modified to pass along the atom type
names as a vtkStringArray named 'atom_types' which are needed by the
new protein ribbon filter.

Change-Id: I2c912d3d8020acbc09f7955f2bf0391cabc52a24