Deprecate SetTickPositions/SetTickLabels for new API
Providing only one method in which both tick positions and labels are
set allows us to ensure that the tick labels are in a consistent state.
Deprecate the old API in favor of this approach, and remove the public
CustomTickLabels API that was never in a release.

Originally I used a two parameter form of SetTickPositions, but that
with deprecation seemed to confuse the Python wrapping. Adding Custom to
the name got rid of the issue, and also makes the VTK deprecation
clearer.

Change-Id: I9726d095915f6a107818772520cfb3bed19c088f