ENH: Full coverage for SparseArray
Extend the unit test for SparseArray to cover all code that can be
reached in normal usage.

NOTE: Some of the const methods cannot be covered because they need to
be applied in a class' const methods.

The original code coverage was 71%. This includes all tests that
exercised SparseArray. But this number is misleading since SparseArray
is template code. Code that is not instantiated is not reported by
code coverage.

For example, these methods were not considered as code by coverage:
GetUniqueCoordinates
GetCoordinateStorage

Also, none of the error conditions were covered by the tests. The unit
test now exercises the error codes and validates the error messages.

The coverage was verified using:
../VTK/Utilities/Maintenance/computeCodeCoverageLocallyForOneTest.sh Common/Core/ -V -R TestArrayAPISparse

Change-Id: I24168f9b157541e28efdfeca7caadfad6fe7b239