Fix and enhance the protein ribbon filter.
Add descriptions to vtkProteinRibbonFilter parameters.
Add a parameter to enable the small molecules visualization.
Those small molecules are now correctly represented (the atom type
fetched from the PDB file and then used was not the atom element
but was related to the atom and residue nomenclature described
in the Chemical Component Dictionary).
vtkMoleculeReader now exports the atom_type array containing the
element atomic number minus 1.
The test now uses a new PDB file.

Change-Id: I38207446635afbcdc573e3b27ca3745f885ea57a