new functionality for vtkTreeHeatmapItem: collapsible subtrees
The user can now collapse and expand subtrees in the dendrogram
by double clicking on a vertex.

I updated TestTreeHeatmapItem to cover this new functionality.

This change required me to create a new option for vtkPruneTreeFilter.
You can now tell this filter to prune the children of the specified
vertex, rather than the vertex itself.  I also added a test for
vtkPruneTreeFilter.

Change-Id: Iaeb9f3eed54f59b9df8c00fca9aab7f538e32c4b