Removing random in regression tests.
It was noticed in skinOrder.tcl that when run under ctest,
math.Random() produces different results to that when run
interactively.
These are the rgb values when skinOrder.tcl is run as a test:
 0.87595 0.598302 0.656878
 0.641134 0.536594 0.537889 ...

When we run the tcl version as as standalone:
 0.504606 0.909988 0.673732
 0.915603 0.533519 0.857481 ...

In the case of Python code the standalone values are the same
as when it is run under ctest.

To resolve this discrepancy, preset values are used.

TestParametricFunctions.tcl has been moved to
Common/ComputationalGeometry/Testing/Tcl but not activated
as a test because the Python one should be sufficient.

Change-Id: I1a35ff34ba139ce5afd15ad3f3c81b4fe444c4aa
Signed-off-by: Andrew Maclean <andrew.amaclean@gmail.com>