Fix potential problem with stream tracers for multiblocks.
Doing a special function to interpolate the point data
from the input to the output
it makes sure the array exists in the input before trying to copy it to the
output. if it doesn't exist in the input but is in the output then we
remove it from the output instead of having bad values there.
this is meant for multiblock data sets where the grids may not have the
same point data arrays or have them in different orders.

Change-Id: Ieab40a7d33ef47d14780cdd8e194da42eabaf7e2