Added getters for various data members and a wrapper hint.
I exposed several data members through getters in order
to be able to develop custom readers in Python. Also added a
hint to export GetUpdateExtent() to Python. In this particular
case, I am combining the XML reader and the MPIImageReader
to read a structured grid. The XML reads the meta data, the
MPIImageReader reads the appended data using MPI/IO.

Change-Id: I7877ecc68f8b200f56ea258eca3bf871feccaaf5