Remove duplicate file name listings in CMakeLists files
This was causing "duplicate class names" in the list of wrapped
classes while working on an "add C# wrappers" topic branch.

The IO/Infovis/CMakeLists.txt file seems to be the only one
that was adding .h files in this context. This commit makes
it consistent with all the other CMakeLists in VTK.

vtkExtractArray was just mistakenly listed twice...

Change-Id: I1dec5e720570a0607f068c732fdbb04037e74795