Fix orientation for some tetra and pyramid contour cases
This issue is originally reported by Paul Edwards.

Contour triangles for all cell types should be oriented consistently
with respect to cell gradients, but they were not. If we adopt the
convention that the triangle orientation should give normal direction
that is opposite to the gradient direction---in the sense that their
dot product is negative, then all tetra cases and some pyramid cases
are incorrect, as revealed by TestContourCases.py. The test
exhaustively sets binary values to cell vertices, contour the cells,
and check the contour triangle normals against cell gradients.

We want to note here that the consistency check here is reversed from
what was originally used by Paul, who considers dot(normal,
gradient)>0 to be correct. The reason for opting for the reverse is to
minimize the code changes---in particular, to avoid changing the case
table for hexahedron.

The fix is simply to fix the first pairs of vertices in the offending
cases.

This changes also include changing an existing test contourCells.py.
This is because penta and hex prism internally use tetra to produce
contours.

Change-Id: I48ef614e2b243ca36338c32674034400e5024ccd