Set VTK_PYTHON_WRAPPED correctly.
Commit b26e66f9 introduced a bug where VTK_PYTHON_WRAPPED global property wasn't
being set correctly. It included modules that weren't being Python wrapped due
to the newly added variables. This caused static builds to fail miserably.
Fixing the logic and also cleanin up the setting of those global properties
closer to the code that actually does the wrapping to avoid such oversight  in
future.

Change-Id: I9884820680f496f2e22ddb287f345e7ff606451e