Add bool IntializeFromCurrentContext() method to vtkRenderWindow.
This new method allows to associate a vtkRenderWindow to an existing
and currently activated OpenGLContext. The window and associated ids
are fetched by analyzing the context properties.
For now, the implementation has been successfully tested on
Windows & Linux.
This new method is used to initialize a RenderWindow from a Java SWT
GLCanvas component.

Change-Id: Ia08b7df58a964a93c265d89ebe7b1316e605ee10