Prevent capturing selection buffer for the cached image
This fixes a bug in the cached-image functionality of the QVTKWidget
class. When performing a selection occasionaly the selection buffer
will be captured and stored as the cached image. This leads to strange
rendering when switching away from the render view.

To fix this we now check if the render window has any active selector
objects before saving the buffer to the cached image.

This fixes ParaView bug #13689.

Change-Id: Ie6b29f30270a67c5f14958e25b5e56c4594413a9