vtkThreshold: Add continuous range option
Previous vtkThreshold considers the scalars from each cell as a
discrete set. A new option is added that considers the scalars as a
continuous interval instead, where the lower bound is the minimum
scalar value and the upper bound is the maximum. This has the effect
of producing more cells when AllScalar is set to be 0: A narrow
threshold might not contain any scalar from the vertices, but it can
still intersect the scalar interval.

Change-Id: I32253e6f4c1ac892010b38ecdcad17e013b133b9