Add simple benchmark to test new text rendering classes.
The benchmark is in Rendering/FreeType/Testing/Cxx, and is disabled
by default since it takes a little while to run and will never actually
fail.

The benchmark compares the new FTTools implementation to FTUtilities,
which was previously the only way to render multi-line text. A variety
of font sizes, orientations, etc are used in order to force cache misses
in the FreeType cache. The test string is unusually long for VTK (500
lines), but this was necessary to ensure that each render occurred on a
measurable timescale. vtkTextRenderer is also tested to benchmark the
overhead that the TextRenderer layer adds.

On my machine in a release-mode build, I get:

FTTools: Total: 23.06  Min: 0.00  Max: 1.29  Ave: 0.23
TextRen: Total: 22.94  Min: 0.00  Max: 1.28  Ave: 0.22
FTUtils: Total: 27.00  Min: 0.01  Max: 1.48  Ave: 0.27

So some of the renderings in the new class are still occurring faster
than clock() can detect, but there is nevertheless an improvement
over FTUtils. It would be useful to figure out a way to benchmark small
strings of 1-5 lines in a reasonable way.

Change-Id: I8f37e3130ed766932f671f315ce790d1772e2410