Support aggregation extracting selection over time.
Instead of breaking a selection into a separate time-history
table for each (block,ID)-tuple, you may call
ReportStatisticsOnly(). Then a single table per
block of the input dataset will report the minimum, maximum,
quartiles, and (for numerical arrays) the average and standard
deviation of the selection over time.

This adds dependencies on the FiltersGeneral and
FiltersStatistics modules to the FiltersExtraction module.

Change-Id: I42c2ef3314a3a4497bd50e1c0dd6cc9fde8ce5fb