Introduce the new AWT and SWT java components.
These 2 classes are intended to replace the old vtkCanvas/vtkPanel
AWT classes at a long term.
The (optional, default is OFF, because it adds a dependency to
Eclipse SWT jar file) SWT component is a pure SWT components
to avoid the AWT-SWT bridge which may generate some difficult to
avoid crashes.
FindSWT provides a CMake procedure to search for the SWT jar file
which stands in the Eclipse plugins directory.

Remove the @Override annotations to avoid issues with OpenJDK.

Change-Id: I96a80b8fdd4c56e591026a17df59e13a6a17ad72