BUG 13817: Fix issues when num-pts in > VTK_CELL_SIZE.
When vtkBoxClipDataSet was applied to dataset with cells have large number of
points per cell, e.g. output of a streamline filter, it would end up corrupting
memory. Changing the code to ensure temporary arrays created to collect the
points in a cell are large enogh, no matter the number of points in the cell.

Change-Id: Idd6b542f9f2d6544ea94d14a973da431f03db6ee