BUG: Fix issues when num-pts in > VTK_CELL_SIZE.
When vtkBoxClipDataSet was applied to dataset with cells have large number of
points per cell, e.g. output of a streamline filter, it would end up corrupting
memory. Changing the code to ensure temporary arrays created to collect the
points in a cell are large enogh, no matter the number of points in the cell.

Change-Id: I033ae945bdcb531e3dbbc75456c56f5f27c38226