Generalizing module config files importing logic.
For applications based on VTK, they themselves might add additional modules. In
that case the module config files locating code needs to look in multiple
locations, those for VTK's modules and those for the extra modules added by the
project. To enable that, we extend vtkModuleAPI to look for the config files in
the VTK_MODULES_DIR as well as any location in the CMAKE_MODULE_PATH. Now, such
applications can simply extend the CMAKE_MODULE_PATH and the module-config files
can be easily imported.

Change-Id: I1b94c96512eae6822e366df06b1e3cdbd363c790