COMP: Python tests missing -D argument
These python tests are missing the -D argument. The macros
add_test_python and add_test_python1 only generate the -D argument if
a second argument is provided to the macro. Usually, the second
argment is the regression baseline directory. Although these tests do
not have a baseline inage, they do require access to VTK_DATA_ROOT.

This patch adds a second argument to the macro, which causes the
python test macros to generate a proper -D argument for the test.

NOTE: The content of the argument is not used. Its presence alone
causes the macros to generate a proper -D argument for the test.

Change-Id: Iec5ccf52a8124212bddb1cfc4cac0f43f6d0f5d8