Replace vtkFreeType(Tools|Utilities) with vtkTextRenderer.
This patch replaces most of the remaining calls to the vtkFreeType
text rendering classes with equivalents that go through the
vtkTextRenderer layer. Most of these were used to calculate bounding
boxes for text actors, which would be incorrect if the text actor was
rendering math text.

Also removed a few unneeded headers and mentions of vtkFreeType.

Change-Id: Iddf7d84dbf46c30774baf96d32edd07b6da77031