COMP: Regression test failures on some platforms
Some windows and linux systems fail the regression tests by minor
amounts. This patch specifies a larger image tolerance.

In particular, the VTK continuous build has a recurring regression
failure for one of these tests. With the increased tolerance, the test
passes.

Tolerances were set so that the tests pass on a Fedora/Mesa system and
were verified on a Fedora/OpenGL system.

Change-Id: Ibf59e01f67e171050411b19ca70041b7b478a1fe