Correct incorrectly calculating min and max cell for each piece.
Previously we incorrectly calculated the min and max cell
for each piece if we had less cells that pieces. This was
caused by using integer division instead of float division.

Also I correct the issue that if we have zero cells we would
divide by zero causing VTK to crash.

Change-Id: Id7f99dfa0b6b85727f326a38c62ada16af8a51ac