Adding support to use MPI based tests in Testing/External
There were a bunch of MPI tests failing when tests were built using
Testing/External. This was because MPI support was not brought in.

In VTK this support was brought in throught code in
Parallel/MPI/CMakeLists.txt. and since vtkMPIParallel was the root
modules for all MPI based modules it ensured that the environment was
brought in for every MPI based modules.

The MPI testing support logic was separated into another file named
vtkTestingMPISupport.cmake which was included at appropriate locations.

Change-Id: I2fa4cb023b77eb9aeb5c818e3625263b616ca42b