Fix bug in squarify treemap layout strategy
The uninitialized bounds array causes all sorts of problems
if zero-sized nodes are reached, since it will leave
uninitialized memory around and leads to erratic layouts
which can be different on different program executions.
The bounds are now initialized to zero to avoid this problem.

Also initializing the point locations to be offscreen,
so labels for zero-sized bins will be drawn offscreen instead
of in the lower left corner of the treemap.

Change-Id: I780b4f4df8936c8f733fb76892bf0b1f1bbf1f79