Fix valgrind error in vtkBlueObeliskDataParser
This fixes "Conditional jump or move depends on uninitialised
value" error issued by valgrind. The atof() call resulted in the
valgrind error when the data is non-numeric. This fixes the error
by using std::stringstream to parse the data instead of atof().

This should fix most of the valgrind errors in the chemistry tests.

Change-Id: I9dfbcf74524d8862d09ded5ba1a85d5dc9d160b3