Eliminated (undefined) negation of INT_MIN
clang ubsan warned running vtkRenderingAnnotationCxx-TestCubeAxes2DMode:

Rendering/Annotation/vtkCubeAxesActor.cxx:1315:30: runtime error: negation of -2147483648 cannot be represented in type 'int'

This happens when 'range' is exactly 0.0, and so log10() gives -infinity,
which when cast to int (apparently) becomes INT_MIN which cannot
be negated due to the asymmetry of two's compliment representation.

Now, check log10() result with newly-created vtkMath::IsFinite API.
If the result is not finite, return 0 from vtkCubeAxesActor::Digits(),
which I'm not sure is the correct thing to return.

Change-Id: I36fc832df98ccb96c1fdc951c42e7e2c6de9d4b4