COMP: Fix static library dependencies on VS 7
Visual Studio 7 and 7.1 do not support direct dependencies between static
library targets cleanly.  CMake works around the problem by adding a
custom "_UTILITY" target in between, but only does it for explicit
add_dependencies and not for target_link_libraries because the extra
targets are ugly and the dependencies are not needed only for compiling
sources.

Now that VTK uses custom commands to run vtkWrapHierarchy when building
module libraries we need the library target ordering dependencies to be
enforced.  In the case of STATIC libraries on VS 7, use an explicit
add_dependencies call to tell CMake to use the work around to enforce the
target build order.  In other cases the existing target_link_libraries is
enough.

Change-Id: I62c934669cedfdb23f7cf8f6cd1e7ee1d1f1d427