Fixed a bug in reading vtu data with polyhedron cells
Paraview bug: http://www.paraview.org/Bug/view.php?id=13892
If there are polydhedron cells in multiple pieces of a vtu file, the
vtkXMLUnstructuredDataReader will fail to read properly the polyhedron
cell and faces info, except the cell in first piece. The bug is in the
logic of figuring out the offset of faces and points when there are
multiple pieces containing polyhedron cells.

Change-Id: Ic1100bc816c1e0f51cc45bc34d26ec92799bd36f