better support for vtkCompositeDataSet output
Allow the vtkCompositeDataSet to contain of any type of vtkDataObject.
Previously, this class assumed that all of the components of the
vtkCompositeDataSet were vtkDataSets.

In order to properly handle vtkCompositeDataSet output, I also had to
add new code within RequestData() to initialize the components of the
output.  Without this change, the output would merely contain the
components of the input vtkCompositeDataSet.

Change-Id: If9c3a83910af99095362919a50d29c6bb3b6db0f