Another approach to fix this degenerate case
This commit resolves the error that were previously experiencing
when a tree was pruned down to a single leaf.  It does so without
changing the color lookup table for normal, collapsed leaves.  This
means that we do not have to update the baseline image to keep the
AutoCollapse test passing.

Change-Id: Idf4fc9e2f9ad2046f298623208769138c1bad841