Add script to upload content-linked external data files to MIDAS
Port the Utilities/Maintenance/ArchiveTestingData.py from ITK to VTK.

Add the VTK Community on the Kitware Midas server as a resource for
downloading VTK ExternalData.  This improves robustness against failed
connections with the Apache file server, protection against the other
server going down, and a web interface to examine and download the
individual files.

Add a script to upload all files corresponding to the content links in
the source tree to the Midas server.  Make the paths of the files mirror
the content link location within the source tree.  Obtain the files from
a local ExternalData object store.

Change-Id: I2dc9234d3ab0499dc2196e68d2c967d9445e10d7