Started on new reader and writer.
libXdmf does the IO

To test, turn on the vtkIOXdmf3 module and python wrapping,
then in python import vtk and connect a vtk.vtkXdmf3Writer to
to anything, give it a filename and tell it to write. It should
create empty xdmf and h5 files where you specified.

Likewise instantiate a vtk.vtkXdmf3Reader, give it the filename
of an xdmf file and call update. It should print out the hierarchy
of XdmfItems in the file.

Common and dataset specific routines are in vtkXdfm3Common.

Common contains a set of static functions, organized in classes. There is
one class per atomic dataset/grid type. They have a 'to' and 'from' method
where the 'to' method is broken up into light (geometry and topology) and heavy
pieces so that RequestInformation and RequestData can do only what they must.
An additional 'Common' class holds functions that are used by the others.

Reader and writer are top level entitities that interface with the
executive and dole out datasets to populate.
Reader produces valid image data, rect grid, struct grid,
unstruct grid and graphs

Have to call xdmf3array::read() to ensure contents are loaded into memory.
Added ifdefd out deepcopy option in process of debugging, and kept it
since it might be useful.

Change-Id: I85964eba08f6f92523e385a19f117840604e51cf