Add translation of GridCollections to Multiblock into reader.
Multiblock and temporal are now working
Still have to enable parallel sieving.
Writer does not yet now what to do when it is given a multiblock.

The large ARL dataset can now be read and written, but translation
errors in the atomic data sets make it nonsensical. Those will
be fixed in testing.

Change-Id: Icadc284a91a97f1d4ec0cdd6ffa2e6a92ad5424e