Add render mode in which values are drawn as recoverable 24bit colors.
With this you can pick an array component and draw that directly
on the screen. Afterward, if you know the min and max, you read pixel
values back and get a hold of the value at each pixel.

TODO:
* Lots of type conversion here, and in the end only using 24 bits
of resolution. When and if we need more accuracy, flesh out
the extensible conversion operator part.
* Surface selection is based on the same concept (ident buffer)
but that was written before VTK has extensible rendering (painters
and passes). Eventually selection should be reimplemented to use
this.
* Would be great to demonstrate render into background followed by
mouse overs.

Change-Id: I44608759278c37c8250f957d596f3f4149382bbb