first light where vtk is actually controlling something
started to fill in code were the scenegraph has state in it.
doing that by looking at the Get methods in doxygen and making
copies of the ones that make sense

started to fill in the actual process of asking ospray to render
something. Having trouble conceptualizing the right traversal
pattern. A simple top down probably won't work here or anywhere
else.