toward full spread of render options
Roint and edge modes could work, but I'm doing something wrong
in how I call ospray (or it has a bug).
Several cases of coloration more or less work now.
TODO: not happy about having to call MapScalars and having to add
the friend class to access the resulting color arrays.
TODO: need to split up the rendering code in ActorNode once I come
to grips with more of what it does.